pyproject.toml
[tool.poetry]
name = "kyc"
version = "0.1.0"
description = ""
authors = ["ademola <ademola@quickcheck.ng>"]

[tool.poetry.dependencies]
python = "^3.8"
redis = "^4.1.2"
django-redis = "^5.2.0"
celery = "^5.2.3"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
psycopg2-binary = "^2.9.3"
django-phonenumber-field = "^6.0.0"
phonenumbers = "^8.12.42"
python-dateutil = "^2.8.2"
factory-boy = "^3.2.1"
pytest = "^6.2.5"
mock = "^4.0.3"
responses = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
